+++ flutter_audio_games/lib/src/widgets/audio_game_menu/audio_game_menu.dart
@@ -18,6 +18,7 @@ class AudioGameMenu extends StatelessWidget {
     this.activateItemSound,
     this.musicFadeInTime,
     this.musicFadeOutTime,
+    this.textStyle = const TextStyle(fontSize: 20),
     super.key,
   });
 
@@ -61,6 +62,9 @@ class AudioGameMenu extends StatelessWidget {
   /// is popped.
   final Duration? musicFadeOutTime;
 
+  /// The text style to use.
+  final TextStyle? textStyle;
+
   /// Build the widget.
   @override
   Widget build(final BuildContext context) {
@@ -83,6 +87,7 @@ class AudioGameMenu extends StatelessWidget {
           selectSound: selectItemSound,
           activateSound: activateItemSound,
           autofocus: index == 0,
+          textStyle: textStyle,
         ),
         itemCount: menuItems.length,
         shrinkWrap: true,