+++ flutter_audio_games/lib/src/widgets/scenes/cut_scene.dart
@@ -14,11 +14,9 @@ class CutScene extends StatelessWidget {
     required this.duration,
     required this.builder,
     this.sound,
+    this.textStyle = const TextStyle(fontSize: 20),
     super.key,
   });
-
-  /// The sound to play for the scene.
-  final Sound? sound;
 
   /// The title of the scene.
   final String sceneTitle;
@@ -32,6 +30,12 @@ class CutScene extends StatelessWidget {
   /// The widget to build after the scene has played out.
   final WidgetBuilder builder;
 
+  /// The sound to play for the scene.
+  final Sound? sound;
+
+  /// The text style to use.
+  final TextStyle? textStyle;
+
   /// Build the widget.
   @override
   Widget build(final BuildContext context) => TimedBuilders(
@@ -41,9 +45,12 @@ class CutScene extends StatelessWidget {
             final s = sound;
             final scaffold = SimpleScaffold(
               title: sceneTitle,
-              body: CenterText(
-                text: sceneText,
+              body: Focus(
                 autofocus: true,
+                child: Text(
+                  sceneText,
+                  style: textStyle,
+                ),
               ),
             );
             if (s == null) {