+++ flutter_audio_games/lib/src/widgets/touch/touch_menu/touch_menu.dart
@@ -24,6 +24,7 @@ class TouchMenu extends StatefulWidget {
     this.downShortcut = GameShortcutsShortcut.arrowDown,
     this.upShortcut = GameShortcutsShortcut.arrowUp,
     this.activateShortcut = GameShortcutsShortcut.enter,
+    this.textStyle = const TextStyle(fontSize: 20),
     super.key,
   });
 
@@ -89,6 +90,9 @@ class TouchMenu extends StatefulWidget {
   /// [GameShortcutsShortcut.space].
   final GameShortcutsShortcut activateShortcut;
 
+  /// The text style to use.
+  final TextStyle? textStyle;
+
   /// Create state for this widget.
   @override
   TouchMenuState createState() => TouchMenuState();
@@ -142,7 +146,7 @@ class TouchMenuState extends State<TouchMenu> {
         .map(
           (final menuItem) => Text(
             menuItem.title,
-            style: const TextStyle(fontSize: 20),
+            style: widget.textStyle,
           ),
         )
         .toList();