+++ flutter_audio_games/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Stop automatically disposing of sounds.
 - Allow load mode to be set on a per-sound basis.
 - Add extension methods to allow calling soloud methods on `AudioSource`s and `SoundHandle`s.
+- Allow setting the text style for `Text`s used throughout the package.
 
 ## 0.23.2